requirements.txt
fastapi
uvicorn
python-dotenv
#pymongo
pymongo[srv]
pydantic
schedule
numpy
pandas
scikit-learn